Create GameManager class to manage playing consecutive games
