+++ lib/controller.rb
@@ -5,33 +5,22 @@ class Controller
   end
 
   def new_game
+    main_game
+    end_of_game
+  end
+
+  def main_game
     until @game.over?
       @display.display_board
       @display.show_current_player(@game.current_player.mark)
       move = @display.ask_for_move
       @game.play_move(move)
     end
-    end_of_game
   end
 
   def end_of_game
     outcome = @game.tie_or_won
     @display.display_board
     @display.show_game_outcome(outcome)
-  end
-
-  def play_again?
-    choice = @display.ask_play_again
-    choice == 'Y'
-  end
-
-  def play
-    play_game = true
-    while play_game
-      new_game
-      play_game = play_again?
-      @game.reset_game
-    end
-    @display.show_exit_message
   end
 end
